Split migration runners and move startup checks into Postgres runner
`scripts/migrate.ts` is split into `scripts/migrate.pg.ts` and
`scripts/migrate.sqlite.ts`:

- `scripts/migrate.pg.ts` retains the legacy chain, baseline assertions,
  and session advisory lock from ADR-0001, and now takes over the database
  readiness wait, `CREATE DATABASE` check, and schema permission probe
  previously performed by `docker-entrypoint.sh` via psql.
- `scripts/migrate.sqlite.ts` runs the Drizzle migrator for SQLite using
  `drizzle-orm/libsql/migrator` over `migrations/sqlite`.
- `scripts/migrate.ts` dispatches dynamically to either runner based on the
  scheme of `DATABASE_URL`.
- Both migration runners are compiled to standalone ESM bundles in `dist/`
  (`dist/migrate.pg.js` and `dist/migrate.sqlite.js`), allowing Docker production
  images to drop TypeScript and tsx.

Verified: `npm run db:verify-upgrade` passes all 7 checks; a fresh Postgres
container successfully boots from scratch through `dist/migrate.pg.js`; all 288
tests pass on both Postgres and SQLite.
