+++ package.json
@@ -5,9 +5,11 @@
   "license": "MIT",
   "scripts": {
     "dev": "tsx server/index.ts",
-    "build": "vite build && npm run build:server:pg && npm run build:server:sqlite",
+    "build": "vite build && npm run build:server:pg && npm run build:server:sqlite && npm run build:migrate:pg && npm run build:migrate:sqlite",
     "build:server:pg": "esbuild server/index.ts --platform=node --packages=external --bundle --format=esm --outfile=dist/index.pg.js",
     "build:server:sqlite": "esbuild server/index.ts --platform=node --packages=external --bundle --format=esm --outfile=dist/index.sqlite.js --tsconfig=tsconfig.sqlite.json",
+    "build:migrate:pg": "esbuild scripts/migrate.pg.ts --platform=node --packages=external --bundle --format=esm --outfile=dist/migrate.pg.js",
+    "build:migrate:sqlite": "esbuild scripts/migrate.sqlite.ts --platform=node --packages=external --bundle --format=esm --outfile=dist/migrate.sqlite.js --tsconfig=tsconfig.sqlite.json",
     "vite:build": "vite build",
     "start": "NODE_ENV=production node dist/index.pg.js",
     "check": "tsc",
