+++ server/db.sqlite.ts
@@ -2,10 +2,21 @@ import { createClient } from "@libsql/client";
 import { drizzle } from "drizzle-orm/libsql";
 import * as schema from "@shared/schema";
 
+import fs from "node:fs";
+import path from "node:path";
+
 if (!process.env.DATABASE_URL) {
   throw new Error(
     "DATABASE_URL must be set. Did you forget to provision a database?",
   );
+}
+
+if (process.env.DATABASE_URL.startsWith("file:") || process.env.DATABASE_URL.startsWith("sqlite:")) {
+  const filePath = process.env.DATABASE_URL.replace(/^(file|sqlite):\/\//, "/").replace(/^(file|sqlite):/, "");
+  const dir = path.dirname(filePath);
+  if (dir && !fs.existsSync(dir)) {
+    fs.mkdirSync(dir, { recursive: true });
+  }
 }
 
 export const client = createClient({ url: process.env.DATABASE_URL });
